refactor: renamed useInfiniteAgentsList to useAgentsList
