+++ apps/meteor/client/components/Omnichannel/hooks/useAgentsList.ts
@@ -1,10 +1,6 @@
 import { useEndpoint } from '@rocket.chat/ui-contexts';
-import { useCallback, useState } from 'react';
+import { useInfiniteQuery } from '@tanstack/react-query';
 import { useTranslation } from 'react-i18next';
-
-import { useScrollableRecordList } from '../../../hooks/lists/useScrollableRecordList';
-import { useComponentDidUpdate } from '../../../hooks/useComponentDidUpdate';
-import { RecordList } from '../../../lib/lists/RecordList';
 
 type AgentsListOptions = {
 	text: string;
@@ -13,56 +9,48 @@ type AgentsListOptions = {
 	excludeId?: string;
 	showIdleAgents?: boolean;
 	onlyAvailable?: boolean;
+	limit?: number;
 };
 
-type AgentOption = { value: string; label: string; _updatedAt: Date; _id: string };
+type AgentOption = {
+	_id: string;
+	value: string;
+	label: string;
+};
 
-export const useAgentsList = (
-	options: AgentsListOptions,
-): {
-	itemsList: RecordList<AgentOption>;
-	initialItemCount: number;
-	reload: () => void;
-	loadMoreItems: (start: number, end: number) => void;
-} => {
+const DEFAULT_QUERY_LIMIT = 25;
+
+export const useAgentsList = (options: AgentsListOptions) => {
 	const { t } = useTranslation();
-	const [itemsList, setItemsList] = useState(() => new RecordList<AgentOption>());
-	const reload = useCallback(() => setItemsList(new RecordList<AgentOption>()), []);
-
 	const getAgents = useEndpoint('GET', '/v1/livechat/users/agent');
-	const { text, onlyAvailable = false, showIdleAgents = true, excludeId, haveAll, haveNoAgentsSelectedOption } = options;
+	const { text, onlyAvailable = false, showIdleAgents = true, excludeId, haveAll, haveNoAgentsSelectedOption, limit = 25 } = options;
 
-	useComponentDidUpdate(() => {
-		options && reload();
-	}, [options, reload]);
-
-	const fetchData = useCallback(
-		async (start: number, end: number) => {
-			const { users: agents, total } = await getAgents({
+	return useInfiniteQuery({
+		queryKey: ['/v1/livechat/users/agent', { text, onlyAvailable, showIdleAgents, excludeId, haveAll, haveNoAgentsSelectedOption }],
+		queryFn: async ({ pageParam: offset = 0 }) => {
+			return getAgents({
 				...(text && { text }),
 				...(excludeId && { excludeId }),
 				showIdleAgents,
 				onlyAvailable,
-				offset: start,
-				count: end + start,
+				offset,
+				count: limit ?? DEFAULT_QUERY_LIMIT,
 				sort: `{ "name": 1 }`,
 			});
-
-			const items = agents.map<AgentOption>((agent) => {
-				const agentOption = {
-					_updatedAt: new Date(agent._updatedAt),
+		},
+		select: (data) => {
+			const items = data.pages.flatMap<AgentOption>(({ users }) => {
+				return users.map((agent) => ({
 					label: `${agent.name || agent._id} (@${agent.username})`,
 					value: agent._id,
 					_id: agent._id,
-				};
-				return agentOption;
+				}));
 			});
 
 			haveAll &&
 				items.unshift({
 					label: t('All'),
 					value: 'all',
-					_updatedAt: new Date(),
 					_id: 'all',
 				});
 
@@ -70,24 +58,19 @@ export const useAgentsList = (
 				items.unshift({
 					label: t('Empty_no_agent_selected'),
 					value: 'no-agent-selected',
-					_updatedAt: new Date(),
 					_id: 'no-agent-selected',
 				});
 
-			return {
-				items,
-				itemCount: total,
-			};
+			return items;
 		},
-		[excludeId, getAgents, haveAll, haveNoAgentsSelectedOption, onlyAvailable, showIdleAgents, t, text],
-	);
-
-	const { loadMoreItems, initialItemCount } = useScrollableRecordList(itemsList, fetchData, 25);
-
-	return {
-		reload,
-		itemsList,
-		loadMoreItems,
-		initialItemCount,
-	};
+		initialPageParam: 0,
+		getNextPageParam: (lastPage) => {
+			const offset = lastPage.offset + lastPage.count;
+			return offset < lastPage.total ? offset : undefined;
+		},
+		initialData: () => ({
+			pages: [{ users: [], total: 0, offset: 0, count: limit ?? DEFAULT_QUERY_LIMIT }],
+			pageParams: [0],
+		}),
+	});
 };
